Revert "Restructured project organization"
This reverts commit cea2a083c9178af96f1131ee50e7ccd4deaa6ff1.
