+++ SimpleDB/IDatabaseRepository.cs
