--- src/Chirp.CLI/Cheep.cs
@@ -1,12 +1,0 @@
-﻿namespace Chirp.CLI;
-
-public class Cheep
-{
-    public string Author
-    {
-        get;
-        set;
-    }
-    public string Message { get; set; }
-    public long Timestamp { get; set; }
-}
