--- src/Chirp.CLI/Chirp.CLI.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-   <ItemGroup>
-        <ProjectReference Include="..\SimpleDB\SimpleDB.csproj" />
-   </ItemGroup>
-    
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net9.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="CsvHelper" Version="33.1.0" />
-  </ItemGroup>
-
-</Project>
